Add run script phase
to copy canonical APCCMSSupport.h from AppCore github repo before compiling.

Also added $(BUILT_PRODUCTS_DIR)/AppCore.framework/Headers to the front of the header search path so it will use the header from whatever version of AppCore you're actually building against, regardless. ...If, that is, you're building CMSSupport as a subproject of an app project that also include AppCore, and not building it as a standalone framework. In the latter case it will use the downloaded one.